+++ source/B2B.Transactions.Api/Configuration/Middleware/Authentication/Bearer/BearerAuthenticationMiddleware.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Text;
 using System.Threading.Tasks;
-using B2B.Transactions.Infrastructure.Authentication;
+using B2B.Transactions.Infrastructure.Configuration.Authentication;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
 using Microsoft.Extensions.Logging;
