+++ source/B2B.Transactions.Api/Configuration/Middleware/Authentication/Bearer/ClaimsEnrichmentMiddleware.cs
@@ -17,7 +17,7 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using B2B.Transactions.Configuration.DataAccess;
-using B2B.Transactions.Infrastructure.Authentication;
+using B2B.Transactions.Infrastructure.Configuration.Authentication;
 using Dapper;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
