+++ source/B2B.Transactions.Api/Configuration/Middleware/Authentication/MarketActors/MarketActorAuthenticatorMiddleware.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Threading.Tasks;
 using B2B.Transactions.Configuration.Authentication;
-using B2B.Transactions.Infrastructure.Authentication;
+using B2B.Transactions.Infrastructure.Configuration.Authentication;
 using B2B.Transactions.Infrastructure.Serialization;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
