+++ source/B2B.Transactions.Infrastructure/Configuration/Authentication/JwtTokenParser.cs
@@ -16,11 +16,10 @@ using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Net.Http.Headers;
-using B2B.CimMessageAdapter;
-using B2B.Transactions.Infrastructure.Authentication.Errors;
+using B2B.Transactions.Infrastructure.Configuration.Authentication.Errors;
 using Microsoft.IdentityModel.Tokens;
 
-namespace B2B.Transactions.Infrastructure.Authentication
+namespace B2B.Transactions.Infrastructure.Configuration.Authentication
 {
     public class JwtTokenParser
     {
