+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -22,8 +22,8 @@ using B2B.Transactions.Configuration;
 using B2B.Transactions.Configuration.Authentication;
 using B2B.Transactions.Configuration.DataAccess;
 using B2B.Transactions.IncomingMessages;
-using B2B.Transactions.Infrastructure.Authentication;
 using B2B.Transactions.Infrastructure.Common;
+using B2B.Transactions.Infrastructure.Configuration.Authentication;
 using B2B.Transactions.Infrastructure.DataAccess;
 using B2B.Transactions.Infrastructure.DataAccess.Transaction;
 using B2B.Transactions.Infrastructure.InternalCommands;
